.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
torch
torchvision
Pillow
google-generativeai
